Move deserializeStripeObject to StripeObject
